requirements.txt
reportlab==4.0.9
pillow==10.1.0
pyyaml==6.0.1
click==8.1.7
jinja2==3.1.2
